membership level

+++ ParkShark.API/Controllers/Members/DTO/MembershipLevelDTO.cs
@@ -10,7 +10,7 @@ namespace ParkShark.API.Controllers.Members.DTO
         public string Membership { get; set; }
         public string Name { get; set; }
         public decimal MonthlyCost { get; set; }
-        public float PSAPriceReductionPercentage { get; set; }
+        public double PSAPriceReductionPercentage { get; set; }
         public TimeSpan PSAMaxTimeInHours { get; set; }
     }
 }
